# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server calculatorstore server.  You can find out\
    \ more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,\
    \ #swagger](http://swagger.io/irc/).      For this sample, you can use the api\
    \ key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger calculator"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "calculator.swagger.io"
basePath: "/v2"
schemes:
- "https"
- "http"
paths:
  /car-economy:
    get:
      tags:
      - "CarEconomy"
      summary: "get all CarEconomy"
      description: "return all CarEconomy items"
      operationId: "getCarEconomy"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CarEconomy"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CarEconomy"
    post:
      tags:
      - "CarEconomy"
      summary: "Add a new CarEconomy"
      description: ""
      operationId: "addPet"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CarEconomy Object added to the database"
        required: true
        schema:
          $ref: "#/definitions/CarEconomy"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "CarEconomy"
    put:
      tags:
      - "CarEconomy"
      summary: "Update an existing CarEconomy"
      description: ""
      operationId: "updatePet"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Pet object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/CarEconomy"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "CarEconomy not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "CarEconomy"
    delete:
      tags:
      - "CarEconomy"
      summary: "Find CarEconomy by id and delete"
      description: "delete a CarEconomy item"
      operationId: "findcalculatorsByTags"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CarEconomy"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "CarEconomy"
  /car-economy/{id}:
    get:
      tags:
      - "CarEconomy"
      summary: "Finds CarEconomy by id"
      description: "get CarEconomy by id"
      operationId: "findCarEconomyByid"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of CarEconomy to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/CarEconomy"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CarEconomy"
...
  /city-params:
    get:
      tags:
      - "CityParams"
      summary: "get all CityParams"
      description: "return all CityParams items"
      operationId: "getCityParams"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParams"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParams"
    post:
      tags:
      - "CityParams"
      summary: "Add a new CityParams"
      description: ""
      operationId: "addCityParams"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParams object that needs to be added in the database"
        required: true
        schema:
          $ref: "#/definitions/CityParams"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "CityParams"
    put:
      tags:
      - "CityParams"
      summary: "Update an existing CityParams"
      description: ""
      operationId: "updateCityParams"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParams object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/CityParams"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "CityParams not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "CityParams"
    delete:
      tags:
      - "CityParams"
      summary: "Find CityParams by id and delete"
      description: "delete a CityParams"
      operationId: "deleteCityParams"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParams"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "CityParams"
  /city-params/{id}:
    get:
      tags:
      - "CityParams"
      summary: "Finds CityParams by id"
      description: "get CityParams by id"
      operationId: "findCityParamsByid"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of CityParams to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParams"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParams"
  /city-params-pop:
    get:
      tags:
      - "CityParamsPop"
      summary: "get all CityParamsPop"
      description: "return all CityParamsPop items"
      operationId: "getCityParamsPop"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsPop"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParamsPop"
    post:
      tags:
      - "CityParamsPop"
      summary: "Add a new CityParamsPop"
      description: ""
      operationId: "addCityParamsPop"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParamsPop object that needs to be added in the database"
        required: true
        schema:
          $ref: "#/definitions/CityParamsPop"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "CityParamsPop"
    put:
      tags:
      - "CityParamsPop"
      summary: "Update an existing CityParamsPop"
      description: ""
      operationId: "updateCityParamsPop"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParamsPop object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/CityParamsPop"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "CityParamsPop not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "CityParamsPop"
    delete:
      tags:
      - "CityParamsPop"
      summary: "Find CityParamsPop by id and delete"
      description: "delete a CityParamsPop"
      operationId: "deleteCityParamsPop"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsPop"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "CityParamsPop"
  /city-params-pop/{id}:
    get:
      tags:
      - "CityParamsPop"
      summary: "Finds CityParamsPop by id"
      description: "get CityParamsPop by id"
      operationId: "findCityParamsPopByid"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of CityParamsPop to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsPop"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParamsPop"
  /city-params-uber:
    get:
      tags:
      - "CityParamsUber"
      summary: "get all CityParamsUber"
      description: "return all CityParamsUber items"
      operationId: "getCityParamsUber"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsUber"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParamsUber"
    post:
      tags:
      - "CityParamsUber"
      summary: "Add a new CityParamsUber"
      description: ""
      operationId: "addCityParamsUber"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParamsUber object that needs to be added in the database"
        required: true
        schema:
          $ref: "#/definitions/CityParamsUber"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "CityParamsUber"
    put:
      tags:
      - "CityParamsUber"
      summary: "Update an existing CityParamsUber"
      description: ""
      operationId: "updateCityParamsUber"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "CityParamsUber object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/CityParamsUber"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "CityParamsUber not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "CityParamsUber"
    delete:
      tags:
      - "CityParamsUber"
      summary: "Find CityParamsUber by id and delete"
      description: "delete a CityParamsUber"
      operationId: "deleteCityParamsUber"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsUber"
        400:
          description: "Invalid tag value"
      x-swagger-router-controller: "CityParamsUber"
  /city-params-uber/{id}:
    get:
      tags:
      - "CityParamsUber"
      summary: "Finds CityParamsUber by id"
      description: "get CityParamsUber by id"
      operationId: "findCityParamsUberByid"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of CityParamsUber to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CityParamsUber"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "CityParamsUber"
...
securityDefinitions:
  calculator_auth:
    type: "oauth2"
    authorizationUrl: "http://calculators.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:calculators: "modify calculators in your account"
      read:calculators: "read your calculators"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  CarEconomy:
    type: "object"
    required:
    - "CRLV"
    - "DPVAT"
    - "IPVA"
    - "carWash"
    - "cityParams"
    - "depreciation"
    - "maintenceCost"
    - "poupanca"
    properties:
      id:
        type: "integer"
      maintenceCost:
        type: "integer"
        format: "int64"
      IPVA:
        type: "integer"
        format: "int32"
      depreciation:
        type: "integer"
        format: "int32"
      poupanca:
        type: "integer"
        format: "int32"
      DPVAT:
        type: "integer"
        format: "int32"
      CRLV:
        type: "integer"
        format: "int32"
      carWash:
        type: "integer"
        format: "int32"
      cityParams:
        type: "integer"
    example:
      IPVA: 1
      poupanca: 5
      cityParams: 3
      CRLV: 7
      maintenceCost: 6
      id: 0
      carWash: 9
      depreciation: 5
      DPVAT: 2
    xml:
      name: "CarEconomy"
  CityParams:
    type: "object"
    required:
    - "baseTax"
    - "cityId"
    - "citySpeed"
    - "cityTax"
    - "distanceTax"
    - "dynamicPrice"
    - "publicTranpPass"
    - "superKMTax"
    - "timeTax"
    - "triggerSuperKM"
    properties:
      id:
        type: "integer"
      baseTax:
        type: "integer"
      distanceTax:
        type: "integer"
      timeTax:
        type: "integer"
      cityTax:
        type: "integer"
      triggerSuperKM:
        type: "integer"
      superKMTax:
        type: "integer"
      citySpeed:
        type: "integer"
      publicTranpPass:
        type: "integer"
      dynamicPrice:
        type: "integer"
      cityId:
        type: "string"
    example:
      timeTax: 5
      cityTax: 5
      triggerSuperKM: 2
      superKMTax: 7
      baseTax: 6
      distanceTax: 1
      citySpeed: 9
      id: 0
      publicTranpPass: 3
      cityId: "cityId"
      dynamicPrice: 2
    xml:
      name: "CityParams"
...
  CityParamsPop:
    type: "object"
    required:
    - "ASP"
    - "GNVPrize"
    - "PAX"
    - "TPH"
    - "alcoholPrize"
    - "cityID"
    - "gasolinePrize"
    - "travelDistance"
    - "travelTime"
    properties:
      id:
        type: "integer"
      TPH:
        type: "integer"
      PAX:
        type: "integer"
      travelDistance:
        type: "integer"
      travelTime:
        type: "integer"
      ASP:
        type: "integer"
      GNVPrize:
        type: "integer"
      alcoholPrize:
        type: "integer"
      gasolinePrize:
        type: "integer"
      cityID:
        type: "string"
    example:
      alcoholPrize: 9
      gasolinePrize: 3
      travelTime: 5
      GNVPrize: 7
      PAX: 1
      travelDistance: 5
      id: 0
      cityID: "cityID"
      TPH: 6
      ASP: 2
    xml:
      name: "CityParamsPop"
  CityParamsUber:
    type: "object"
    required:
    - "ASP"
    - "GNVPrize"
    - "PAX"
    - "TPH"
    - "alcoholPrize"
    - "cityID"
    - "gasolinePrize"
    - "travelDistance"
    - "travelTime"
    properties:
      id:
        type: "integer"
      TPH:
        type: "integer"
      PAX:
        type: "integer"
      travelDistance:
        type: "integer"
      travelTime:
        type: "integer"
      ASP:
        type: "integer"
      GNVPrize:
        type: "integer"
      alcoholPrize:
        type: "integer"
      gasolinePrize:
        type: "integer"
      cityID:
        type: "string"
    example:
      alcoholPrize: 9
      gasolinePrize: 3
      travelTime: 5
      GNVPrize: 7
      PAX: 1
      travelDistance: 5
      id: 0
      cityID: "cityID"
      TPH: 6
      ASP: 2
    xml:
      name: "CityParamsUber"
...